Restructuring and removing scanning.
